+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import TaskBoard from "components/TaskBoard";
+import Board from "components/Board";
 import { Container, Stack, Title } from "@mantine/core";
 
 export default function Home() {
@@ -6,9 +6,9 @@ export default function Home() {
         <Container h="95vh" size="xl" mt="xl">
             <Stack h="100%" gap="xs">
                 <Title order={2} ta="center">
-                    To-do list
+                    ( To-do list )
                 </Title>
-                <TaskBoard />
+                <Board />
             </Stack>
         </Container>
     );