+++ src/types/board.types.ts
@@ -1,8 +1,8 @@
 export type ColumnId = "PRIORITY" | "IN_PROGRESS" | "ON_HOLD" | "NOTE";
 
-export type Task = {
+export type Item = {
     id: string;
     content: string;
 };
 
-export type Board = Record<ColumnId, Task[]>;
+export type Board = Record<ColumnId, Item[]>;